NMS-17948 add review changs
